TodoApi/requirements.txt
asgiref==3.6.0
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
django-redis==5.2.0
django-environ==0.10.0
psycopg2==2.9.6